Add Cxx-Qt-Macro crate
This allows us to add other items to cxx-qt, other than proc-macros.
e.g. a QtObject trait or similar.

It might even allow us to move CxxQtThread into cxx-qt, instead of
cxx-qt-lib.
