+++ crates/cxx-qt-macro/Cargo.toml
@@ -1,0 +1,28 @@
+# SPDX-FileCopyrightText: 2023 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
+# SPDX-FileContributor: Andrew Hayzen <andrew.hayzen@kdab.com>
+# SPDX-FileContributor: Gerhard de Clercq <gerhard.declercq@kdab.com>
+# SPDX-FileContributor: Leon Matthes <leon.matthes@kdab.com>
+#
+# SPDX-License-Identifier: MIT OR Apache-2.0
+[package]
+name = "cxx-qt-macro"
+version.workspace = true
+authors = ["Andrew Hayzen <andrew.hayzen@kdab.com>","Leon Matthes <leon.matthes@kdab.com>", "Gerhard de Clercq <gerhard.declercq@kdab.com>"]
+edition.workspace = true
+license.workspace = true
+description = "A set of macros for Qt/C++ interop in Rust"
+repository.workspace = true
+homepage = "https://kdab.github.io/cxx-qt/book/"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+cxx-qt-gen.workspace = true
+proc-macro2.workspace = true
+syn.workspace = true
+
+[dev-dependencies]
+cxx.workspace = true
+cxx-qt-lib.workspace = true
+cxx-qt.workspace = true
